### 2023-02-15 #5154 @editorialbot
[PRE REVIEW]: Regl-Scatterplot: A Scalable Interactive JavaScript-based Scatter Plot Library
**Submitting author:** @flekschas (<a href="http://orcid.org/0000-0001-8432-4835">Fritz Lekschas</a>)
**Repository:** https://github.com/flekschas/regl-scatterplot
**Branch with paper.md** (empty if default branch):
**Version:** v1.5.1
