### 2023-02-15 #5154 @danielskatz
👋 @flekschas - thanks for your submission. While it looks useful in research, the editors will now discuss if it meets the [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) criterion for review by JOSS. You should hear back in a week or two.

In the meantime, you could work on the possibly missing DOI that editorialbot suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
