+++ paper/refs.bib
@@ -112,7 +112,7 @@
 
 @misc{gotreescape,
   author = {Li, Guozheng},
-  title = {GoTreeScape: Navigate and Explore the Tree Visualization Design Space},
+  title = {GoTreeScape Prototype System},
   year = {2022},
   url = {https://github.com/bitvis2021/GoTreeScape},
 }
