### 2023-02-15 #5154 @flekschas
Thank you for the heads up @danielskatz. If it helps, my thinking is that `regl-scatterplot` "extracts knowledge from large data sets" by making sure that large data sets can be efficiently visualized and interactively explored as a scatter plot.
