### 2023-02-15 ff71771
Add info panel linking to GeoNames
