+++ assets/transition.fe3346b8.js
@@ -1,1 +1,0 @@
-import{b as S,c as g,a as h,s as F,d as v}from"./utils.10233b37.js";import{t as E}from"./apache-arrow.3ccd3ece.js";import"./vendor.cfd24b35.js";const L=["#FFFF00","#1CFFD9","#FF34FF","#FF4A46","#008941","#1966FF","#C00069","#FFDBE5","#FF9900","#8148D5"],q=fetch("https://storage.googleapis.com/flekschas/regl-scatterplot/cities.arrow");S("Loading...");const k=document.querySelector("#canvas"),w=document.querySelector("#num-points"),z=document.querySelector("#num-points-value"),u=document.querySelector("#point-size"),C=document.querySelector("#point-size-value"),A=document.querySelector("#opacity"),f=document.querySelector("#opacity-value"),d=document.querySelector("#click-lasso-initiator"),I=document.querySelector("#reset"),P=document.querySelector("#export"),p=document.querySelector("#example-transition");p.setAttribute("class","active");p.removeAttribute("href");let s=1.5;const e=g({canvas:k,pointSize:s,lassoOnLongPress:!0});h(e);P.addEventListener("click",()=>F(e));console.log(`Scatterplot v${e.get("version")}`);w.disabled=!0;A.disabled=!0;const m=t=>{s=t,u.value=s,C.innerHTML=s,e.set({pointSize:s})},x=t=>m(+t.target.value);u.addEventListener("input",x);const H=t=>{e.set({lassoInitiator:t.target.checked})};d.addEventListener("change",H);d.checked=e.get("lassoInitiator");const b=()=>{e.reset()};I.addEventListener("click",b);m(1.5);q.then(t=>E(t)).then(t=>{v();const n=t.data[0].children.map(o=>o.values),c=n[n.length-2],i=n[n.length-1],a=new Float32Array(c.length);for(let o=0;o<c.length;o++)a[o]=1;z.innerHTML=t.numRows,f.innerHTML="Automatic",e.draw({x:n[0],y:n[1],z:i,w:c}),e.set({colorBy:"z",pointColor:L,opacityBy:"w",opacity:Array.from({length:10},(o,y)=>.33+y/9*.33)});let r=0;const l=()=>{r++;const o=r%3*2;e.draw({x:n[o],y:n[o+1],z:i,w:[c,a,a][r%3]},{transition:!0,transitionDuration:750,transitionEasing:"quadInOut"}).then(()=>{setTimeout(l,2500)})};setTimeout(l,2500)});
