+++ assets/transition.dba4da1a.js
@@ -1,0 +1,3 @@
+import{b as y,c as g,a as S,s as F,d as v}from"./utils.10233b37.js";import{t as E}from"./apache-arrow.3ccd3ece.js";import"./vendor.cfd24b35.js";const L=["#FFFF00","#1CFFD9","#FF34FF","#FF4A46","#008941","#1966FF","#C00069","#FFDBE5","#FF9900","#8148D5"],q=fetch("https://storage.googleapis.com/flekschas/regl-scatterplot/cities.arrow");y("Loading...");const w=document.querySelector("#footer");w.classList.remove("hidden");const f=document.querySelector("#info-content");f.innerHTML=`
+<p><a href="https://www.geonames.org/about.html" target="_blank">GeoNames \u2013 Cities Dataset</a> visualized in three: by the cities' geographic location, by the total population across contintents, and by the citie's latitude distribution.</p>
+`;const b=document.querySelector("#canvas"),k=document.querySelector("#num-points"),z=document.querySelector("#num-points-value"),u=document.querySelector("#point-size"),C=document.querySelector("#point-size-value"),A=document.querySelector("#opacity"),H=document.querySelector("#opacity-value"),d=document.querySelector("#click-lasso-initiator"),I=document.querySelector("#reset"),P=document.querySelector("#export"),p=document.querySelector("#example-transition");p.setAttribute("class","active");p.removeAttribute("href");let s=1.5;const e=g({canvas:b,pointSize:s,lassoOnLongPress:!0});S(e);P.addEventListener("click",()=>F(e));console.log(`Scatterplot v${e.get("version")}`);k.disabled=!0;A.disabled=!0;const m=t=>{s=t,u.value=s,C.innerHTML=s,e.set({pointSize:s})},x=t=>m(+t.target.value);u.addEventListener("input",x);const D=t=>{e.set({lassoInitiator:t.target.checked})};d.addEventListener("change",D);d.checked=e.get("lassoInitiator");const M=()=>{e.reset()};I.addEventListener("click",M);m(1.5);q.then(t=>E(t)).then(t=>{v();const n=t.data[0].children.map(o=>o.values),a=n[n.length-2],r=n[n.length-1],i=new Float32Array(a.length);for(let o=0;o<a.length;o++)i[o]=1;z.innerHTML=t.numRows,H.innerHTML="Automatic",e.draw({x:n[0],y:n[1],z:r,w:a}),e.set({colorBy:"z",pointColor:L,opacityBy:"w",opacity:Array.from({length:10},(o,h)=>.33+h/9*.33)});let c=0;const l=()=>{c++;const o=c%3*2;e.draw({x:n[o],y:n[o+1],z:r,w:[a,i,i][c%3]},{transition:!0,transitionDuration:750,transitionEasing:"quadInOut"}).then(()=>{setTimeout(l,2500)})};setTimeout(l,2500)});
