### 2023-02-28 #5154 @fabian-s
:wave: @fil & @aokad,

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
