+++ src/index.js
@@ -283,6 +283,8 @@ const createScatterplot = (
   let selection = [];
   const selectionSet = new Set();
   const selectionConnecionSet = new Set();
+  let filteredPoints = [];
+  const filteredPointSet = new Set();
   let numPoints = 0;
   let numPointsInView = 0;
   let lassoActive = false;
@@ -515,6 +517,19 @@ const createScatterplot = (
     return computedPointSizeMouseDetection * pointScale * pxNdc * 0.66;
   };
 
+  const getPoints = () => {
+    if (filteredPointSet.size > 0)
+      return searchIndex.points.filter((_, i) => filteredPointSet.has(i));
+    return searchIndex.points;
+  };
+
+  const getPointsInBBox = (x0, y0, x1, y1) => {
+    const pointsInBBox = searchIndex.range(x0, y0, x1, y1);
+    if (filteredPointSet.size > 0)
+      return pointsInBBox.filter((i) => filteredPointSet.has(i));
+    return pointsInBBox;
+  };
+
   const raycast = () => {
     const [xGl, yGl] = getMouseGlPos();
     const [xNdc, yNdc] = getScatterGlPos(xGl, yGl);
@@ -522,7 +537,7 @@ const createScatterplot = (
     const pointSizeNdc = getPointSizeNdc();
 
     // Get all points within a close range
-    const pointsInBBox = searchIndex.range(
+    const pointsInBBox = getPointsInBBox(
       xNdc - pointSizeNdc,
       yNdc - pointSizeNdc,
       xNdc + pointSizeNdc,
@@ -556,7 +571,7 @@ const createScatterplot = (
     // get the bounding box of the lasso selection...
     const bBox = getBBox(lassoPolygon);
     // ...to efficiently preselect potentially selected points
-    const pointsInBBox = searchIndex.range(...bBox);
+    const pointsInBBox = getPointsInBBox(...bBox);
     // next we test each point in the bounding box if it is in the polygon too
     const pointsInPolygon = [];
     pointsInBBox.forEach((pointIdx) => {
@@ -638,6 +653,9 @@ const createScatterplot = (
     Math.floor(index / stateTexRes) / stateTexRes + stateTexEps,
   ];
 
+  const isPointFilteredOut = (pointIdx) =>
+    filteredPointSet.size > 0 && !filteredPointSet.has(pointIdx);
+
   const deselect = ({ preventEvent = false } = {}) => {
     if (lassoClearEvent === LASSO_CLEAR_ON_DESELECT) lassoClear();
     if (selection.length) {
@@ -675,15 +693,21 @@ const createScatterplot = (
     for (let i = selection.length - 1; i >= 0; i--) {
       const pointIdx = selection[i];
 
-      if (pointIdx < 0 || pointIdx >= numPoints) {
+      if (
+        pointIdx < 0 ||
+        pointIdx >= numPoints ||
+        isPointFilteredOut(pointIdx)
+      ) {
         // Remove invalid selection
         selection.splice(i, 1);
-      } else {
-        selectionSet.add(pointIdx);
-        const texCoords = indexToStateTexCoord(pointIdx);
-        selectionBuffer.push(texCoords[0]);
-        selectionBuffer.push(texCoords[1]);
+        continue;
       }
+
+      selectionSet.add(pointIdx);
+      selectionBuffer.push.apply(
+        selectionBuffer,
+        indexToStateTexCoord(pointIdx)
+      );
     }
 
     selectedPointsIndexBuffer({
@@ -697,6 +721,51 @@ const createScatterplot = (
     if (!preventEvent) pubSub.publish('select', { points: selection });
 
     draw = true;
+  };
+
+  /**
+   * @param {number | number[]} point
+   * @param {import('./types').ScatterplotMethodOptions['hover']} options
+   */
[…]
