### 2023-03-03 2b3cb25
Document filter functions
