### 2023-03-03 91aa396
Expose selected and filtered point indices
