+++ src/index.js
@@ -280,11 +280,11 @@ const createScatterplot = (
   let mouseDownTime = null;
   let mouseDownPosition = [0, 0];
   let mouseDownTimeout = -1;
-  let selection = [];
-  const selectionSet = new Set();
-  const selectionConnecionSet = new Set();
+  let selectedPoints = [];
+  const selectedPointsSet = new Set();
+  const selectedPointsConnectionSet = new Set();
   let filteredPoints = [];
-  const filteredPointSet = new Set();
+  const filteredPointsSet = new Set();
   let numPoints = 0;
   let numPointsInView = 0;
   let lassoActive = false;
@@ -518,15 +518,15 @@ const createScatterplot = (
   };
 
   const getPoints = () => {
-    if (filteredPointSet.size > 0)
-      return searchIndex.points.filter((_, i) => filteredPointSet.has(i));
+    if (filteredPointsSet.size > 0)
+      return searchIndex.points.filter((_, i) => filteredPointsSet.has(i));
     return searchIndex.points;
   };
 
   const getPointsInBBox = (x0, y0, x1, y1) => {
     const pointsInBBox = searchIndex.range(x0, y0, x1, y1);
-    if (filteredPointSet.size > 0)
-      return pointsInBBox.filter((i) => filteredPointSet.has(i));
+    if (filteredPointsSet.size > 0)
+      return pointsInBBox.filter((i) => filteredPointsSet.has(i));
     return pointsInBBox;
   };
 
@@ -600,7 +600,7 @@ const createScatterplot = (
     const isNormal = stateIndex === 0;
     const lineIdCacher =
       stateIndex === 1
-        ? (lineId) => selectionConnecionSet.add(lineId)
+        ? (lineId) => selectedPointsConnectionSet.add(lineId)
         : identity;
 
     // Get line IDs
@@ -619,7 +619,7 @@ const createScatterplot = (
     const buffer = pointConnections.getData().opacities;
 
     lineIds
-      .filter((lineId) => !selectionConnecionSet.has(+lineId))
+      .filter((lineId) => !selectedPointsConnectionSet.has(+lineId))
       .forEach((lineId) => {
         const index = pointConnectionMap[lineId][0];
         const numPointPerLine = pointConnectionMap[lineId][2];
@@ -654,16 +654,16 @@ const createScatterplot = (
   ];
 
   const isPointFilteredOut = (pointIdx) =>
-    filteredPointSet.size > 0 && !filteredPointSet.has(pointIdx);
+    filteredPointsSet.size > 0 && !filteredPointsSet.has(pointIdx);
 
   const deselect = ({ preventEvent = false } = {}) => {
     if (lassoClearEvent === LASSO_CLEAR_ON_DESELECT) lassoClear();
-    if (selection.length) {
+    if (selectedPoints.length) {
       if (!preventEvent) pubSub.publish('deselect');
-      selectionConnecionSet.clear();
-      setPointConnectionColorState(selection, 0);
-      selection = [];
-      selectionSet.clear();
+      selectedPointsConnectionSet.clear();
+      setPointConnectionColorState(selectedPoints, 0);
+      selectedPoints = [];
+      selectedPointsSet.clear();
       draw = true;
     }
   };
@@ -677,35 +677,35 @@ const createScatterplot = (
     const pointIdxsArr = Array.isArray(pointIdxs) ? pointIdxs : [pointIdxs];
 
     if (merge) {
-      selection = unionIntegers(selection, pointIdxsArr);
+      selectedPoints = unionIntegers(selectedPoints, pointIdxsArr);
     } else {
       // Unset previously highlight point connections
-      if (selection && selection.length)
-        setPointConnectionColorState(selection, 0);
-      selection = pointIdxsArr;
+      if (selectedPoints && selectedPoints.length)
[…]
