+++ README.md
@@ -607,7 +607,9 @@ can be read and written via [`scatterplot.get()`](#scatterplot.get) and [`scatte
 | opacity                               | float                                        | `1`                                 | Must be in ]0, 1]                                               | `true`   | `false`     |
 | opacityInactiveMax                    | float                                        | `1`                                 | Must be in [0, 1]                                               | `true`   | `false`     |
 | opacityInactiveScale                  | float                                        | `1`                                 | Must be in [0, 1]                                               | `true`   | `false`     |
-| points                                | tuple                                        | `[0.5, 2.3]`                        |                                                                 | `false`  | `false`     |
+| points                                | tuple[]                                      | `[[0.5, 2.3], ...]`                 |                                                                 | `false`  | `false`     |
+| selectedPoints                        | int[]                                        | `[4, 2]`                            |                                                                 | `false`  | `false`     |
+| filteredPoints                        | int[]                                        | `[4, 2]`                            |                                                                 | `false`  | `false`     |
 | pointsInView                          | int[]                                        | `[1, 2, 12]`                        |                                                                 | `false`  | `false`     |
 | pointColor                            | quadruple                                    | `[0.66, 0.66, 0.66, 1]`             | single value or list of hex, rgb, rgba                          | `true`   | `false`     |
 | pointColorActive                      | quadruple                                    | `[0, 0.55, 1, 1]`                   | single value or list of hex, rgb, rgba                          | `true`   | `false`     |
