### 2023-03-04 144097e
Fix an issue with repeated zooming to points
