### 2023-03-20 #5154 @fabian-s
hi @rowlandm @xiaohk @j-andrews7

would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
