### 2023-03-22 flekschas/regl-scatterplot#106 @Fil
JOSS Review
Ref. https://github.com/openjournals/joss-reviews/issues/5275

The only issue I found is:
* community guidelines are missing

_paper_

* It is technically incorrect to say that D3 relies on SVG (it can do canvas too, for instance).
* typos:
  * s/spacial/spatial/
  * "study, end test" s/end/and/

_example_

the [observable example](https://observablehq.com/@flekschas/regl-scatterplot) could be enhanced by adding interaction on selection; for example it could be as simple as adding 2 cells and changing 1:

~~~diff
+ mutable selection = null

+ scatterplot.subscribe("select", ({points}) => mutable selection = points)

- createScatterplot = require('regl-scatterplot@0.7.6/dist/regl-scatterplot.min.js')
+ createScatterplot = require('regl-scatterplot@1')
~~~

(Or it could be rewritten in a more "observable" way by using [viewof and data flow](https://observablehq.com/@fil/hello-regl-scatterplot); see also https://observablehq.com/@radames/hello-regl-scatterplot for earlier work.)
