### 2023-03-23 flekschas/regl-scatterplot#106 @fabian-s
Thanks!
