### 2023-03-23 flekschas/regl-scatterplot#106 @Fil
This gives me more "feature requests", for example a variable pointSize that would depend on the zoom factor (like deepScatter does, I think?). But out of scope of this issue :)
