### 2023-03-24 7de8723
Updated the wording around D3 and SVG reliance
