+++ .github/ISSUE_TEMPLATE/bug.md
@@ -1,0 +1,15 @@
+---
+name: 'Bug report'
+about: 'Report a bug to help us improve'
+labels: 'Bug :bug:'
+---
+
+Oh snap! Looks like something went south. Thank you for taking the time to report the issue.
+
+Before you file it, please:
+
+- [ ] Check for duplicate issues.
+- [ ] Describe the bug: what did you do to trigger it, what was your expected behavior, and what did you experience instead.
+- [ ] Use the latest versions of `regl-scatterplot`.
+- [ ] Provide an [minimal, reproducible example](https://stackoverflow.com/help/minimal-reproducible-example) of the bug.
+- [ ] If applicable, include error messages, screenshots, GIFs, videos, or working examples.
