+++ .github/ISSUE_TEMPLATE/enhancement.md
@@ -1,0 +1,17 @@
+---
+name: 'Enhancement'
+about: 'Suggest a new idea or improvement for `regl-scatterplot`'
+labels: 'Enhancement :tada:'
+---
+
+You have an idea how to enhance `regl-scatterplot`? That's awesome!
+
+Before you file your idea, please:
+
+- [ ] Check for duplicate issues.
+- [ ] Describe the feature's goal, motivating use cases, its expected behavior, and impact.
+- [ ] If you are proposing a new API, please explain how the current APIs fail to support your goal.
+- [ ] If you are proposing to change an existing API, please explain if your proposed is backward compatible or incompatible.
+- [ ] If applicable, include screenshots, GIFs, videos, or working examples.
+
+You are encouraged to prototype multiple alternative APIs or implementation approaches for your proposed feature. Doing so often leads to a better solution.
