+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,18 @@
+> Write one to two sentences summarizing this PR
+
+## Description
+
+> What was changed in this pull request?
+
+> Why is it necessary?
+
+Fixes #___
+
+## Checklist
+
+- [ ] Provided a concise title as a [semantic commit message](https://www.conventionalcommits.org) (e.g. "fix: correctly handle undefined properties")
+- [ ] Updated CHANGELOG.md
+- [ ] Added or updated Tests added or updated
+- [ ] Documentation added or updated in README.md
+- [ ] Example(s) added or updated
+- [ ] Screenshot, gif, or video attached for visual changes
