### 2023-03-24 flekschas/regl-scatterplot#106 @flekschas
Thanks for the review!

Since this is my first JOSS review, apologies in advance as I might not know about the best way to address the review.

I got started and put together [a PR](https://github.com/flekschas/regl-scatterplot/pull/108) to address several issues by:
- Fixing types
- Rephrasing that "D3 relies on SVG"
- Adding community guidelines

Should I go ahead and merge the PR or await your feedback first?

> the [observable example](https://observablehq.com/@flekschas/regl-scatterplot) could be enhanced by adding interaction on selection

I agree! I'll get to that tomorrow. Thanks for the quick examples!

> This gives me more "feature requests", for example a variable pointSize that would depend on the zoom factor (like deepScatter does, I think?). But out of scope for this issue :)

Absolutely! It'd be fun to allow customizing the behavior. The good news is that the point size is already dependent on the zoom level. Just not linearly.
_tl/dr:_ Normally, the point size is either zoom invariant (points are always drawn with a radius in absolute pixels) or bound to the geometric zoom. I found both to not be ideal. With zoom invariant point sizes the scatter gets too sparse upon zooming in. On the other hand, having the point size be bound linearly to the geometric zoom means that zooming in won't reveal finer details for points that superimpose each other as they keep superimposing each other at every zoom level. As a solution, in `regl-scatterplot` if the camera scaling is above `1` (the user zooms in) the point size is scaled by the inverse hyperbolic sine function. If the camera scaling is below `0` the point size is linearly scaled according to the camera scaling.
