### 2023-03-24 flekschas/regl-scatterplot#106 @Fil
Thanks! With this I've checked the last item on my review.
