### 2023-03-24 flekschas/regl-scatterplot#106 @fabian-s
Wonderful, thank you both for your quick and constructive action!
