### 2023-03-24 05a72ca
Joss review 106 (#108)
* Fix typos

* Updated the wording around D3 and SVG reliance

* Add PR and issue templates

* Add contribution guidelines
