+++ CONTRIBUTING.md
@@ -1,0 +1,37 @@
+# Contributing to Regl-Scatterplot
+
+:wave: Welcome and thank you for considering to contribute to `regl-scatterplot`.
+
+In this guide you will get an overview of the contribution workflow from opening an issue, creating a PR, reviewing, and merging the PR.
+
+## Getting Started
+
+To get an overview of the tool, checkout the [demos](https://flekschas.github.io/regl-scatterplot/) and read the [README](README.md). In particular, you might find the following sections helpful:
+
+- [Getting started](https://github.com/flekschas/regl-scatterplot#getting-started)
+- [API](https://github.com/flekschas/regl-scatterplot#api)
+  - [Properties](https://github.com/flekschas/regl-scatterplot#properties)
+  - [Events](https://github.com/flekschas/regl-scatterplot#events)
+- [Trouble shooting](https://github.com/flekschas/regl-scatterplot#trouble-shooting)
+
+## Help! I am lost... :sob:
+
+Head over to our [discussion board](https://github.com/flekschas/regl-scatterplot/discussions) and we're happy to help.
+
+## Reporting Issues :bug: or Feature Requests :bulb:
+
+If you discover a bug in `regl-scatterplot` or have an idea for a new feature, please [submit a new issue](https://github.com/flekschas/regl-scatterplot/issues/new). Make sure the issue is following either the [bug report](/.github/ISSUE_TEMPLATE/bug.md) or [feature request](/.github/ISSUE_TEMPLATE/enhancement.md) template.
+
+For questions on how to use `regl-scatterplot`, please use the [discussion board](https://github.com/flekschas/regl-scatterplot/discussions).
+
+## Contribute Code :nail_care:
+
+If you have a concrete plan how to improve `regl-scatterplot` and have created a [ticket](#reporting-issues-or-feature-requests), send us your pull requests! If you're wondering what a pull request (PR) is, please check out [GitHub's how to](https://help.github.com/articles/using-pull-requests/).
+
+A maintainer will review your pull request. Once the pull request is approved and passes continuous integration checks, we will merge it and add it to the next scheduled release.
+
+#### Pull Request Checklist :white_check_mark:
+
+- Ensure that your pull request answers the relevant questions in the [pull request template](/.github/PULL_REQUEST_TEMPLATE.md).
+- Ensure that your pull requests successfully passes all continuous integration tests with GH Actions.
+- Optionally, if your pull request introduces a new feature, it'd be great to add a new demo to `/example`.
