### 2023-03-29 flekschas/regl-scatterplot#109 @xiaohk
JOSS Review from @xiaohk
Hi @flekschas, I'm honored to [review this JOSS submission](https://github.com/openjournals/joss-reviews/issues/5275). Thank you for making this fantastic tool and writing a great JOSS manuscript!

Everything looks good to me. I saw that the contributing guideline is addressed in #108 👍

I noticed a minor issue is about the reference list of the manuscript.

>**References**: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

The reference list is complete, but it seems some entries have missing information. For example, the paper "Zeno: Interactive framework for machine learning evaluation" is actually published in CHI 2023 (probably the news is after your JOSS submission 😅 ). It'd be great if the `regl-scatterplot` manuscript can include its venue information. (Zeno has a bibtex block on their [README](https://github.com/zeno-ml/zeno#citation)).

```latex
@inproceedings{cabrera23zeno,
  author = {Cabrera, Ángel Alexander and Fu, Erica and Bertucci, Donald and Holstein, Kenneth and Talwalkar, Ameet and Hong, Jason I. and Perer, Adam},
  title = {Zeno: An Interactive Framework for Behavioral Evaluation of Machine Learning},
  year = {2023},
  isbn = {978-1-4503-9421-5/23/04},
  publisher = {Association for Computing Machinery},
  address = {New York, NY, USA},
  url = {https://doi.org/10.1145/3544548.3581268},
  doi = {10.1145/3544548.3581268},
  booktitle = {CHI Conference on Human Factors in Computing Systems},
  location = {Hamburg, Germany},
  series = {CHI '23}
}
```

https://github.com/flekschas/regl-scatterplot/blob/05a72ca675e72983ba152211ec9e1d6969b5bc09/paper/refs.bib#L134-L139

Also, it seems the manuscript cites Visinity twice with different references. I think it can just use the second reference. I suggest double checking the reference list to make sure all entries are correct and up-to-date.

```
Warchol, S. (2023). Visinity: Navigate and explore the tree visualization design space. https://github.com/labsyspharm/Visinity

Warchol, S., Krueger, R., Nirmal, A. J., Gaglia, G., Jessup, J., Ritch, C. C., Hoffer, J., Muhlich, J., Burger, M. L., Jacks, T., Santagata, S., Sorger, P. K., & Pfister, H. (2023). Visinity: Visual spatial neighborhood analysis for multiplexed tissue imaging data. IEEE Transactions on Visualization and Computer Graphics, 29(1), 106–116. https://doi.org/10.1109/TVCG.
2022.3209378
```
