+++ paper/refs.bib
@@ -69,7 +69,7 @@
   address = {Los Alamitos, CA, USA},
 }
 
-@ARTICLE{micallef2017towards,
+@article{micallef2017towards,
   author = {Micallef, Luana and Palmas, Gregorio and Oulasvirta, Antti and Weinkauf, Tino},
   journal = {IEEE Transactions on Visualization and Computer Graphics},
   title = {Towards Perceptual Optimization of the Visual Design of Scatterplots},
@@ -79,7 +79,6 @@
   pages = {1588-1599},
   doi = {10.1109/TVCG.2017.2674978}
 }
-
 
 @mastersthesis{hindersson2021scatterplot,
   title = {Scatterplot Patterns: Using animation to enhance scatterplots},
@@ -124,18 +123,18 @@
   url = {https://github.com/flekschas/jupyter-scatter},
 }
 
-@misc{visinity,
-  author = {Warchol, Simon},
-  title = {Visinity: Navigate and Explore the Tree Visualization Design Space},
+@inproceedings{cabrera23zeno,
+  author = {Cabrera, Ángel Alexander and Fu, Erica and Bertucci, Donald and Holstein, Kenneth and Talwalkar, Ameet and Hong, Jason I. and Perer, Adam},
+  title = {Zeno: An Interactive Framework for Behavioral Evaluation of Machine Learning},
   year = {2023},
-  url = {https://github.com/labsyspharm/Visinity},
-}
-
-@misc{zeno,
-  author = {Cabrera, \'{A}ngel Alexander},
-  title = {Zeno: Interactive framework for machine learning evaluation},
-  year = {2023},
-  url = {https://github.com/zeno-ml/zeno},
+  isbn = {978-1-4503-9421-5/23/04},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/3544548.3581268},
+  doi = {10.1145/3544548.3581268},
+  booktitle = {CHI Conference on Human Factors in Computing Systems},
+  location = {Hamburg, Germany},
+  series = {CHI '23}
 }
 
 @misc{regl,
