### 2023-03-31 flekschas/regl-scatterplot#109 @flekschas
Thanks a lot for the review @xiaohk! 🙏 Regarding the two visinity citations, my thinking was to directly reference the software repository to make it easy for the curious reader to find them but I agree, it's a bit confusing.

I fixed both issues in #110
