### 2023-03-31 flekschas/regl-scatterplot#109 @xiaohk
Got you, thanks!! Great paper!
