### 2023-04-03 flekschas/regl-scatterplot#109 @fabian-s
Wonderful, thank you both for your quick and constructive action!
