### 2023-04-03 #5275 @fabian-s
@flekschas

At this point could you:

- [ ] Fix/remove the invalid DOI for the "Zeno" paper and run "check references" again to verify the change

Then:
