### 2023-04-03 #5275 @flekschas
@fabian-s

I assume that the invalid DOI of the Zeno paper is due to the fact that the paper is just about to get released and doi.org is fairly slow at updating its database. I double checked the DOI (10.1145/3544548.3581268) and it matches the author's bibtex listed at https://zenoml.com/about and the related DOI of their preprint https://arxiv.org/abs/2302.04732.

Shall I move forward removing the DOI anyway?
