### 2023-04-04 #5275 @flekschas
The tagged [version is v1.6.3](https://github.com/flekschas/regl-scatterplot/releases/tag/v1.6.3) and it's archived at [Zenode with doi:10.5281/zenodo.7796642](https://doi.org/10.5281/zenodo.7796642)
