### 2023-04-06 28cf4cb
Fixed Issue#112 - isMouseInCanvas out of sync
Changed isMouseInCanvas from false to true.
