### 2023-04-09 e22ce77
Update refs.bib
Remove isbn numbers as they might cause a parsing error
