+++ paper/refs.bib
@@ -46,7 +46,6 @@
   author = {B\"{a}uerle, Alex and Cabrera, \'{A}ngel Alexander and Hohman, Fred and Maher, Megan and Koski, David and Suau, Xavier and Barik, Titus and Moritz, Dominik},
   title = {Symphony: Composing Interactive Interfaces for Machine Learning},
   year = {2022},
-  isbn = {9781450391573},
   publisher = {Association for Computing Machinery},
   address = {New York, NY, USA},
   url = {https://doi.org/10.1145/3491102.3502102},
@@ -127,7 +126,6 @@
   author = {Cabrera, Ángel Alexander and Fu, Erica and Bertucci, Donald and Holstein, Kenneth and Talwalkar, Ameet and Hong, Jason I. and Perer, Adam},
   title = {Zeno: An Interactive Framework for Behavioral Evaluation of Machine Learning},
   year = {2023},
-  isbn = {978-1-4503-9421-5/23/04},
   publisher = {Association for Computing Machinery},
   address = {New York, NY, USA},
   url = {https://doi.org/10.1145/3544548.3581268},
