### 2023-04-09 #5275 @flekschas
@fabian-s Could you make the editorialbot try it again? I've removed the ISBN numbers from two references, which I suspect might have been incorrect. The authors, title, year, journal, publisher, DOI, etc. should suffice.
