### 2023-04-11 #5275 @fabian-s
EiCs: note that the "invalid" DOI editorialbot complains about is valid but refers to a document that is not yet published / updated in the DOI.org database
