### 2023-04-11 #5275 @danielskatz
@flekschas - there's a references heading missing - https://github.com/flekschas/regl-scatterplot/pull/115 will fix it
