+++ paper/refs.bib
@@ -29,7 +29,7 @@
 }
 
 @article{narechania2022vitality,
-  title = {VITALITY: Promoting Serendipitous Discovery of Academic Literature with Transformers &amp; Visual Analytics},
+  title = {VITALITY: Promoting Serendipitous Discovery of Academic Literature with Transformers \& Visual Analytics},
   author = {Narechania, Arpit and Karduni, Alireza and Wesslen, Ryan and Wall, Emily},
   journal = {IEEE Transactions on Visualization and Computer Graphics},
   year = {2022},
@@ -103,7 +103,7 @@
 
 @misc{eodash,
   author = {Euro Data Cube},
-  title = {EODash: Software behind the RACE dashboard by ESA and the European Commission as well as the Earth Observing Dashboard by NASA, ESA, and JAXA},
+  title = {EODash: Software behind the RACE dashboard by ESA and the {E}uropean {C}ommission as well as the {E}arth {O}bserving {D}ashboard by NASA, ESA, and JAXA},
   year = {2022},
   url = {https://github.com/eurodatacube/eodash},
 }
@@ -117,7 +117,7 @@
 
 @misc{jscatter,
   author = {Lekschas, Fritz},
-  title = {Jupyter-Scatter: An interactive scatter plot widget for Jupyter Notebook, Lab, and Google Colab},
+  title = {Jupyter-{S}catter: An interactive scatter plot widget for {J}upyter {N}otebook, {L}ab, and {G}oogle {C}olab},
   year = {2022},
   url = {https://github.com/flekschas/jupyter-scatter},
 }
@@ -158,7 +158,7 @@
 
 @misc{plotlyjs,
   author = {{Plotly, Inc}},
-  title = {Plotly.js: Open-source JavaScript charting library behind Plotly and Dash},
+  title = {Plotly.js: Open-source JavaScript charting library behind {P}lotly and {D}ash},
   year = {2023},
   url = {https://github.com/plotly/plotly.js},
 }
@@ -246,7 +246,7 @@
 
 @misc{kingma2013auto,
   author = {Kingma, Diederik P and Welling, Max},
-  title = {Auto-Encoding Variational Bayes},
+  title = {Auto-Encoding Variational {B}ayes},
   publisher = {arXiv},
   year = {2013},
   doi = {10.48550/ARXIV.1312.6114},
