+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Lekschas
+  given-names: Fritz
+  orcid: "https://orcid.org/0000-0001-8432-4835"
+doi: 10.5281/zenodo.7796642
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Lekschas
+    given-names: Fritz
+    orcid: "https://orcid.org/0000-0001-8432-4835"
+  date-published: 2023-04-11
+  doi: 10.21105/joss.05275
+  issn: 2475-9066
+  issue: 84
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5275
+  title: "Regl-Scatterplot: A Scalable Interactive JavaScript-based
+    Scatter Plot Library"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05275"
+  volume: 8
+title: "Regl-Scatterplot: A Scalable Interactive JavaScript-based
+  Scatter Plot Library"
