### 2023-04-11 #5275 @flekschas
Thanks everyone! DOI seems to work fine and the PDF appears on JOSS! 🎉🌈🦄💃👻🤘
